Seperate letter counting from MostCommonLetter
Abstracting this logic out into its own plugin means we can reuse it for
future modules e.g. it will be easy to add a least common letter
statistics generator.

This is possibly violating the "you ain't gonna need it" principle, but
as it is easy and quick enough to do, it seems beneficial rather than
trying to do it when additional functionality has been added.